Redirecting from Interview view to the result view

+++ Muqabalati.Service/DTOs/AnswerModel.cs
@@ -9,7 +9,7 @@ namespace Muqabalati.Service.DTOs
     public class AnswerModel
     {
         public string Answer { get; set; } = string.Empty;
-        public float TimeToken { get; set; }
+        public double TimeToken { get; set; }
 
     }
 }
